Code restructured, index.js resolver split to single resolvers
